+++ display.h
@@ -28,7 +28,7 @@ void initDisplay(void);
  *
  * @param percentage The percentage altitude taken from altitude_get() in altitude.c
 */
-void displayPercentage (uint16_t percentage);
+void displayPercentage (uint32_t percentage);
 
 
 /**
